refactor: RequestHandler에 포함된 HTTP 요청 처리 로직을 DispatcherServlet으로 분리
