--- src/main/java/FrontController.java
@@ -1,4 +1,0 @@
-public class FrontController {
-
-
-}
