+++ src/main/java/webserver/RequestHandler.java
@@ -2,12 +2,9 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import application.controller.Controller;
-import application.controller.ControllerResolver;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
 import webserver.utils.HttpConstants;
-import webserver.utils.HttpField;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -17,7 +14,7 @@ import java.net.Socket;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
-    private static final ControllerResolver controllerResolver = ControllerResolver.getInstance();
+    private final DispatcherServlet dispatcherServlet = DispatcherServlet.getInstance();
     private final Socket connection;
 
     public RequestHandler(Socket connectionSocket) {
@@ -32,37 +29,19 @@ public class RequestHandler implements Runnable {
             HttpRequest httpRequest = new HttpRequest(in);
             HttpResponse httpResponse = new HttpResponse();
 
-            processRequest(httpRequest, httpResponse);
+            dispatcherServlet.dispatch(httpRequest, httpResponse);
+
             sendResponse(httpResponse, out);
         } catch (IOException e) {
             logger.error(e.getMessage());
         }
     }
 
-    private void processRequest(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        Controller controller = resolveController(httpRequest);
-        controller.process(httpRequest, httpResponse);
-    }
-
-    private Controller resolveController(HttpRequest httpRequest) {
-        String path = httpRequest.getHeader(HttpField.PATH);
-        String method = httpRequest.getMethod();
-        return controllerResolver.resolve(path, method);
-    }
-
     private void sendResponse(HttpResponse httpResponse, OutputStream out) throws IOException {
-        writeResponseHeaders(httpResponse, out);
-        writeBlankLine(out);
+        out.write(httpResponse.getHeaderBytes());
+        out.write(HttpConstants.CRLF.getBytes());
         writeResponseBody(httpResponse, out);
         out.flush();
-    }
-
-    private void writeResponseHeaders(HttpResponse httpResponse, OutputStream out) throws IOException {
-        out.write(httpResponse.getHeaderBytes());
-    }
-
-    private void writeBlankLine(OutputStream out) throws IOException {
-        out.write(HttpConstants.CRLF.getBytes());
     }
 
     private void writeResponseBody(HttpResponse httpResponse, OutputStream out) throws IOException {
