+++ src/main/java/webserver/utils/HttpField.java
@@ -2,13 +2,11 @@ package webserver.utils;
 
 public final class HttpField {
 
-    public static final String METHOD = "Method";
     public static final String URI = "URI";
     public static final String VERSION = "Version";
     public static final String CONTENT_TYPE = "Content-Type";
     public static final String CONTENT_LENGTH = "Content-Length";
     public static final String AUTHORIZATION = "Authorization";
-    public static final String PATH = "Path";
     public static final String LOCATION = "Location";
 
     private HttpField() {
